+++ http/test/helper.ml
@@ -32,7 +32,7 @@ let with_server ?error_handler ?read_header_timeout handler ~f =
       ~max_accepts_per_batch:64
       ~on_handler_error:`Raise
       Tcp.Where_to_listen.of_port_chosen_by_os
-      ~f:(fun _addr reader writer ->
+      (fun _addr reader writer ->
       let server =
         Shuttle_http.Server.create ?read_header_timeout ?error_handler reader writer
       in