+++ shuttle/lib/output_channel.ml
@@ -3,6 +3,10 @@ open! Async
 include Output_channel0
 module Input_channel = Input_channel0
 
+let write_timeout t = t.write_timeout
+let time_source t = t.time_source
+let buffer_size t = Bytebuffer.capacity t.buf
+
 let of_pipe info pipe_writer =
   Async.Unix.pipe info
   >>| fun (`Reader rd, `Writer wr) ->